backend/pyproject.toml
[tool.poetry]
    description = "Diango rest example app"
    name        = "src"
    repository  = "https://github.com/MyOrg/django-rest-example"

    authors = ["Hoang"]
    classifiers = [
        "Programming Language :: Python :: 3.10.4",
        "Framework :: Django :: 4.0.6",
        "Topic :: Software Development :: Libraries :: Python Modules",
    ]
    license = "BSL-1.0"
    maintainers = ["Hoang"]
    packages = [
        {include = "src", from = "."},
        {include = "poetry_scripts", from = "."},
    ]
    readme = "README.md"
    version = "0.0.1"

    [tool.poetry.scripts]

        # src   = "applyservice.cli:main"

        makemigrations = "poetry_scripts.django:makemigrations"
        migrate        = "poetry_scripts.django:migrate"
        newapp         = "poetry_scripts.django:startapp"
        server         = "poetry_scripts.django:runserver"
        shell          = "poetry_scripts.django:shell"
        sslserver      = "poetry_scripts.django:runsslserver"

    [tool.poetry.dependencies]
        python = "^3.10.5"

        django-axes                 = "5.32"
        django-csp                  = "3.7"
        django-health-check         = "3.16.5"
        django-http-referrer-policy = "1.1.1"
        django-permissions-policy   = "4.9"
        django-split-settings       = "1.1"
        django-stubs-ext            = "0.4"

        bcrypt          = "4.0.1"
        gunicorn        = "20.1"
        psycopg2-binary = "2.9.3"
        python-decouple = "3.6"
        sentry-sdk      = "1.5.12"
        structlog       = "21.5.0"



        bandit      = ">=1.7.2"
        curlify     = ">=2.2.1"
        flake8      = ">=4.0.1"
        ipython     = ">=8.0.1"
        jedi        = ">=0.18.1"
        pep8-naming = ">=0.12.1"
        pre-commit  = ">=2.17.0"
        pyright     = ">=1.1.245"
        virtualenv  = ">=20.13.1"

        django                  = "4.2.6"
        django-coverage-plugin  = "^3.1.0"
        django-debug-toolbar    = "^4.2.0"
        django-extra-checks     = "^0.13.3"
        django-migration-linter = "^5.0.0"
        django-querycount       = "0.8.3"
        django-test-migrations  = "^1.3.0"
        django-types            = "^0.18.0"
        flake8-django           = "^1.4"
        pytest-django           = "^4.5.2"

        flake8-bugbear             = "^22.7.1"
        flake8-docstrings          = "^1.6.0"
        flake8-implicit-str-concat = "^0.3.0"
        flake8-pytest-style        = "^1.6"
        wemake-python-styleguide   = "^0.16.1"

        nitpick        = "^0.32"
        requests-cache = "0.9.7"

        pytest              = "^7.2"
        pytest-cov          = "^4.0"
        pytest-deadfixtures = "^2.2"
        pytest-factoryboy   = ">=2.1.0"
        pytest-randomly     = "^3.11"
        pytest-testmon      = "^1.3"
        pytest-timeout      = "^2.1"

        pytest-html  = ">=3.1.1"
        pytest-mock  = ">=3.6.1"
        pytest-xdist = ">2.4.0"

        faker = ">=19.6.0"

        hypothesis = "^6.47.0"
        nplusone   = "^1.0"

        doc8                     = "^1.1.1"
        tomlkit                  = "^0.11.0"

        dennis        = "^1.1"
        dotenv-linter = "^0.4"
        dump-env      = "^1.3"
        polint        = "^0.4"
        safety        = "^2.3.1"
        yamllint      = "^1.26"
    mysqlclient = "^2.2.0"
    django-structlog = "^6.0.0"
    djangorestframework = "^3.14.0"
    psutil = "^5.9.5"
    html2text = "^2020.1.16"
    jsonschema = "^4.19.1"
    phonenumbers = "^8.13.22"
    urlextract = "^1.8.0"
    lxml = "^4.9.3"
    xmltodict = "^0.13.0"
    pycountry = "^22.3.5"
    better-profanity = "^0.7.0"
    ftfy = "^6.1.1"
    markdown = "^3.4.4"
    charset-normalizer = "^3.3.0"
    django-redis = "^5.4.0"
    tabulate = "^0.9.0"
    django-dirtyfields = "^1.9.2"
    django-hashid-field = "^3.3.7"
    croniter = "^1.4.1"
[tool.nitpick]
    cache = "never"
    style = "./styleguide/nitpick-style.toml"

[tool.pyright]
    exclude                = ["**/node_modules", "**/__pycache__"]
    include                = ["src"]
    reportMissingImports   = true
    reportMissingTypeStubs = false

    pythonPlatform = "Linux"
    pythonVersion  = "3.10"

[build-system]
    build-backend = "poetry.core.masonry.api"
    requires      = ["poetry-core>=1.0.0"]
